+++ src/node/NodeUtils.ts
@@ -1,4 +1,4 @@
-import * as escodegen from 'escodegen';
+import * as escodegen from '@javascript-obfuscator/escodegen';
 import * as estraverse from 'estraverse';
 import * as ESTree from 'estree';